+++ brewtils/rest/system_client.py
@@ -547,6 +547,7 @@ class SystemClient(object):
                 brewtils.plugin.request_context, "current_request", None
             )
             request.has_parent = request.parent is not None
+
             ec = EasyClient(
                 bg_host=brewtils.plugin.CONFIG.bg_host,
                 bg_port=brewtils.plugin.CONFIG.bg_port,
@@ -595,6 +596,13 @@ class SystemClient(object):
         topic = kwargs.pop("_topic", None)
         propagate = kwargs.pop("_propagate", None)
 
+        if parent:
+            requester = getattr(
+                brewtils.plugin.request_context.current_request, "requester", None
+            )
+        else:
+            requester = None
+
         if system_display:
             metadata["system_display_name"] = system_display
         if publish:
@@ -626,6 +634,7 @@ class SystemClient(object):
             parent=parent,
             metadata=metadata,
             parameters=kwargs,
+            requester=requester,
         )
 
         request.parameters = self._resolve_parameters(command, request)